created DeleteAlert class
