+++ app/src/main/java/com/example/bime/fragments/EditEntryFragment.kt
@@ -17,6 +17,7 @@ import com.example.bime.R
 import com.example.bime.classes.CheckCalendar
 import com.example.bime.classes.CheckRadioButtons
 import com.example.bime.classes.CheckTimeInput
+import com.example.bime.classes.DeleteAlert
 import com.google.android.material.bottomappbar.BottomAppBar
 import java.time.LocalDate
 
@@ -94,27 +95,7 @@ class EditEntryFragment : Fragment() {
             goBack()
         }
         deleteButton.setOnClickListener {
-                val dialogBuilder = AlertDialog.Builder(this.activity)
-
-                // set message of alert dialog
-                dialogBuilder.setMessage("Are you sure you want to delete this entry?")
-                    // if the dialog is cancelable
-                    .setCancelable(true)
-                    // positive button text and action
-                    .setPositiveButton("Proceed") { dialog, id ->
-                        onEntryDelete()
-                    }
-                    // negative button text and action
-                    .setNegativeButton("Cancel") { dialog, id ->
-                        dialog.cancel()
-                    }
-
-            // create dialog box
-                val alert = dialogBuilder.create()
-                // set title for alert dialog box
-                alert.setTitle("Delete Entry")
-                // show alert dialog
-                alert.show()
+            var deleteAlert = DeleteAlert(this.activity, this::onEntryDelete)
             }
 
         view.findViewById<ImageView>(R.id.back_arrow).setOnClickListener { goBack() }
